task[docker]: Adds volume for database
This is a named volume (bucket of file storage), completely managed by Docker.

We did:
  <edited docker-compose.yml to add the volume>

  docker-compose stop database
  docker-compose rm -f database
  docker-compose up -d database
  docker-compose exec web bin/rails db:create db:migrate

I expected myapp_development to already exist, but it also said
myapp_test already existed.  This must be due to my previous attempts
and the "magic" of where/how Docker stores this data.

In the end, the data I had created in the last step was still there.