task[webpacker]: Gets the webpack dev server going
We want our Webpacker managed files compiled automatically.  The docker
way to do this is to run the webpack dev server in it's own container.

So, we did:
<edit the docker-compose.yml>
We mounted the local files, so that the dev server can pickup changes

Then:
docker-compose up -d web
docker-compose up -d webpack_dev_server