task[decision]: Choosing to not have a /gems volume
I got to the part of the book that shows a method to not install gems
from scratch and use a volume to get bundle behavior like we're used to.

That means bundle only installs the gems you don't have installed.

I'm choosing to avoid this complexity, because for this project and a
future personal project I think it is clearer to do build and then
install all the gems.  I'm ok with the time trade off (unless my
internet connection sucks).